WIP Close but failing to write to container
Hmmmm... turns out I can't mount EFS in such a way that it is visible to
the container. Either that or the documentation is too deep for me to
have found yet.

Boo. I think I have to ditch doing this with Lambda Docker Containers.

But all is not lost: I recently learned about Lambda Layers and looks
like there's one already built for git. Next up I'm trying that.